4.21

+++ controllers/blogs.js
@@ -41,8 +41,14 @@ blogsRouter.get('/:id', async (request, response) => {
 
 
 blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end()
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  const blog = await Blog.findById(request.params.id)
+  if ( blog.user.toString() === decodedToken.id.toString() ){
+    await Blog.findByIdAndRemove(request.params.id)
+    response.status(204).end()
+  }
+  
+  
 })
 
 blogsRouter.put('/:id', async (request, response) => {
